make `get_player_by_name` async, rename old fn
Currently, the existing commands system is all that requires the
existing blocking version of `get_player_by_name` function. After
discussing in the Discord, it seems renaming the old blocking method
with a suffix of `_blocking` in addition to adding an async version
under the original (unmodified) name of `get_player_by_name`.

This change was required to work on the 'Check if player already joined'
project.